Break component utils into separate modules
