+++ src/js/components/utils/useFetchTasksIfMissing.tsx
@@ -1,0 +1,35 @@
+import { useCallback, useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { RouteComponentProps } from 'react-router-dom';
+
+import { AppState, ThunkDispatch } from '@/store';
+import { fetchObjects } from '@/store/actions';
+import { Project } from '@/store/projects/reducer';
+import { selectTasksByProject } from '@/store/tasks/selectors';
+import { OBJECT_TYPES } from '@/utils/constants';
+
+export default (
+  project: Project | null | undefined,
+  routeProps: RouteComponentProps,
+) => {
+  const dispatch = useDispatch<ThunkDispatch>();
+  const selectTasksWithProps = useCallback(selectTasksByProject, []);
+  const tasks = useSelector((state: AppState) =>
+    selectTasksWithProps(state, routeProps),
+  );
+
+  useEffect(() => {
+    if (project && !tasks) {
+      // Fetch tasks from API
+      dispatch(
+        fetchObjects({
+          objectType: OBJECT_TYPES.TASK,
+          filters: { project: project.id },
+          shouldSubscribeToObject: true, 
+        }),
+      );
+    }
+  }, [dispatch, project, tasks]);
+
+  return { tasks };
+};
